Update ObjectMgr.h
armor is stored in and read from the database as int32

+++ src/server/game/Globals/ObjectMgr.h
@@ -502,7 +502,7 @@ struct PetLevelInfo
     uint16 stats[MAX_STATS];
     uint16 health;
     uint16 mana;
-    uint16 armor;
+    uint32 armor;
     uint16 min_dmg;
     uint16 max_dmg;
 };
